added component testing
